Fix Services

+++ WildPrices.Business/Services/Implementation/ProductService.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using WildPrices.Business.DTOs.ProductDtos;
+using WildPrices.Business.Exceptions;
 using WildPrices.Business.Services.Common;
 using WildPrices.Data.Entities;
 using WildPrices.Data.Repositories.Contracts;
@@ -8,15 +9,15 @@ namespace WildPrices.Business.Services.Implementation
 {
     public class ProductService : IProductService
     {
-        private IMapper _mapper;
-        private IProductRepository _productRepository;
+        private readonly IMapper _mapper;
+        private readonly IProductRepository _productRepository;
 
         public ProductService(IMapper mapper, IProductRepository productRepository)
         {
             _mapper = mapper;
             _productRepository = productRepository;
         }
-        public DesiredPriceDto GetDisiredPriceAsync(double desiredPrice)
+        public DesiredPriceDto GetDisiredPrice(double desiredPrice)
         {
             return new DesiredPriceDto
             {
@@ -25,7 +26,7 @@ namespace WildPrices.Business.Services.Implementation
         }
 
         public async Task CreateProductWhenDoesNotExistAsync(ProductFromWildberriesDto model,
-            DesiredPriceDto desiredPriceDto, MinAndMaxPriceDto minAndMaxPriceDto)
+            DesiredPriceDto desiredPriceDto, string id)
         {
             if (model == null)
             {
@@ -38,10 +39,11 @@ namespace WildPrices.Business.Services.Implementation
                 Name = model.Name,
                 Link = model.Link,
                 Image = model.Image,
-                MaxPrice = minAndMaxPriceDto.MaxPrice,
-                MinPrice = minAndMaxPriceDto.MinPrice,
+                MaxPrice = 0,
+                MinPrice = 0,
                 DesiredPrice = desiredPriceDto.DesiredPrice,
-                IsDesiredPrice = false
+                IsDesiredPrice = false,
+                UserId = id
             };
 
             var product = _mapper.Map<ProductEntity>(productForCreationDto);
@@ -52,6 +54,12 @@ namespace WildPrices.Business.Services.Implementation
         public async Task DeleteProductAsync(int id)
         {
             var entity = await _productRepository.GetByIdAsync(id);
+
+            if(entity == null)
+            {
+                throw new NotFoundException("product");
+            }
+
             await _productRepository.DeleteAsync(entity);
         }
 
@@ -65,9 +73,9 @@ namespace WildPrices.Business.Services.Implementation
             return await _productRepository.GetAllIsNotDesiredAsync();
         }
 
-        public async Task<IEnumerable<ProductEntity>> GetAllProductAsync()
+        public async Task<IEnumerable<ProductEntity>> GetAllProductAsync(string userId)
         {
-            return await _productRepository.GetAllAsync();
+            return await _productRepository.GetAllAsyncByUserId(userId);
         }
 
         public async Task<int> GetProductArticleByIdAsync(int id)
@@ -79,6 +87,22 @@ namespace WildPrices.Business.Services.Implementation
         {
             var product = _mapper.Map<ProductEntity>(model);
             await _productRepository.UpdateAsync(product);
+        }
+
+        public async Task<int> GetProductIdByArticle(int article)
+        {
+            return await _productRepository.GetProductIdByArticle(article);
+        }
+
+        public async Task<ProductEntity> GetProductById(int id)
+        {
+            var product = await _productRepository.GetByIdAsync(id);
+
+            if(product == null)
+            {
+                throw new NotFoundException(nameof(product));
+            }
+            return product;
         }
     }
 }
